refactor AuthDetector into extension functions
